requirements.txt
alembic==1.13.1
black==24.3.0
cachetools==5.3.3
cryptography==42.0.5
email-validator==2.1.1
fastapi>=0.110.0,<0.111.0
httpx==0.27.0
pre-commit==3.7.0
pydantic-settings==2.2.1
pylint==3.1.0
pymysql==1.1.0
pytest==8.1.1
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
sqlalchemy==2.0.29
trio==0.25.0
uvicorn==0.29.0